few improvements
Feel free to ignore these. I just want to put comments. But there is no easy way to comment on a commit that is not part of a PR.

Few other general comments:

1. A reference to the CLINT spec. [https://sifive.cdn.prismic.io/sifive/0d163928-2128-42be-a75a-464df65e04e0_sifive-interrupt-cookbook.pdf]
2. I am not sure if ACLINT can claim full backward compliance with sifive CLINT.
   CLINT spec talks about these things things ACLINT doesn't specify
   - local interrupts
   - Vectored interrupt
   - priorities withing local interrupts
  -  clarification about priority & preemption between MSWI/SSWI

In practice, ACLINT is compatible with CLINT as none of the implementation uses vectored/local interrupt in CLINT.
But we need to clarify that to avoid confusion.